handled float value for boolean and overwrote the transform()
