+++ tests/unittests/test_boolean_conversion.py
@@ -1,5 +1,6 @@
 import unittest
-from tap_google_sheets.singer_transform import transform
+from singer.transform import NO_INTEGER_DATETIME_PARSING
+from tap_google_sheets.streams import new_transform
 
 schema = {
     "properties": {
@@ -36,24 +37,29 @@ metadata = {(): {
         "inclusion": "available"
     }
 }
+
+class MockTransformer():
+    '''Mock Request object'''
+    def __init__(self, integer_datetime_fmt=NO_INTEGER_DATETIME_PARSING, pre_hook=None):
+        self.integer_datetime_fmt = integer_datetime_fmt
+        self.pre_hook = pre_hook
+
 class TestBooleanDataType(unittest.TestCase):
-    def test_string_returned_for_non_boolean_columns(self):
+    def test_string_returned_for_non_string_value(self):
         '''
         Verify that the non boolean values returns string.
         '''
-        record = {
-            "__sdc_row": 1,
-            "Boolean": "string",
-        }
-        transformed_record = transform(record, schema, metadata)
-        self.assertEqual(transformed_record.get("Boolean"), record.get("Boolean"))
+        data = "string"
+        transformer = MockTransformer()
+        transformed_data = new_transform(transformer, data, "boolean", schema, '')
+        print(transformed_data)
+        self.assertEqual(transformed_data[1], "string")
 
-    def test_string_returned_for_boolean_columns(self):
+    def test_boolean_returned_for_boolean_columns(self):
         '''
         Verify that the boolean values in a column returns boolean values.'''
-        record = {
-            "__sdc_row": 1,
-            "Boolean": True,
-        }
-        transformed_record = transform(record, schema, metadata)
-        self.assertEqual(transformed_record.get("Boolean"), record.get("Boolean"))
+        data = True
+        transformer = MockTransformer()
+        transformed_data = new_transform(transformer, data, "boolean", schema, '')
+        print(transformed_data)
+        self.assertEqual(transformed_data[1], True)
